// src/components/spot-trading/CoinPairSwitcher/enum.ts
export enum RateState {
	GREEN = "text-labels whitespace-nowrap sm:text-x-small text-green",
	DEFAULT = "text-labels whitespace-nowrap sm:text-x-small text-neutral-400",
}

export enum ValueState {
	GREEN = "text-labels sm:text-x-small whitespace-nowrap text-green",
	DEFAULT = "text-labels sm:text-x-small whitespace-nowrap text-neutral-400",
}